reorganize toggle component, styles
