+++ apps/consent-ui/src/components/Header/Header.module.scss
@@ -42,3 +42,18 @@
     width: 250px;
   }
 }
+
+.toggle-full {
+  @apply hidden;
+  @media (min-width: theme('screens.md')) {
+    @apply flex;
+  }
+}
+
+.toggle-abbr {
+  @apply uppercase;
+  @apply flex;
+  @media (min-width: theme('screens.md')) {
+    @apply hidden;
+  }
+}
