+++ apps/consent-ui/src/components/Header/index.tsx
@@ -21,7 +21,7 @@ import Image, { StaticImageData } from 'next/image';
 
 import { ValidLanguage, getTranslation } from '@/i18n';
 import { defaultLanguage, supportedLanguages } from '@/i18n/settings';
-import LanguageToggle from '@/components/LanguageToggle';
+import LanguageToggle from '@/components/Header/LanguageToggle';
 import OhcrnImage from '@/public/ohcrn_large.svg';
 
 import styles from './Header.module.scss';
